+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/view/_internal/_structure/_class/ClassStructureNode.kt
@@ -14,7 +14,7 @@ import dev.turingcomplete.intellijbytecodeplugin.bytecode.MethodDeclarationUtils
 import dev.turingcomplete.intellijbytecodeplugin.bytecode.TypeUtils
 import dev.turingcomplete.intellijbytecodeplugin.org.objectweb.asm.Opcodes
 import dev.turingcomplete.intellijbytecodeplugin.org.objectweb.asm.tree.ClassNode
-import dev.turingcomplete.intellijbytecodeplugin.view._internal._structure.SearchProvider
+import dev.turingcomplete.intellijbytecodeplugin.view._internal._structure.GoToProvider
 import dev.turingcomplete.intellijbytecodeplugin.view._internal._structure._common.HtmlTextNode
 import dev.turingcomplete.intellijbytecodeplugin.view._internal._structure._common.HyperLinkNode
 import dev.turingcomplete.intellijbytecodeplugin.view._internal._structure._common.TextNode
@@ -29,7 +29,7 @@ import javax.swing.Icon
 internal class ClassStructureNode(private val classNode: ClassNode, private val classFile: VirtualFile)
   : ValueNode(displayValue = { cxt -> TypeUtils.toReadableName(classNode.name, cxt.typeNameRenderMode) },
               icon = determineClassIcon(classNode),
-              searchProvider = SearchProvider.Class(classNode.name)) {
+              goToProvider = GoToProvider.Class(classNode.name)) {
 
   // -- Companion Object -------------------------------------------------------------------------------------------- //
 
@@ -90,7 +90,7 @@ internal class ClassStructureNode(private val classNode: ClassNode, private val 
       add(ValueNode("Super:",
                     { ctx -> TypeUtils.toReadableName(superName, ctx.typeNameRenderMode) },
                     icon = AllIcons.Hierarchy.Supertypes,
-                    searchProvider = SearchProvider.Class(superName)))
+                    goToProvider = GoToProvider.Class(superName)))
     }
   }
 
@@ -119,7 +119,7 @@ internal class ClassStructureNode(private val classNode: ClassNode, private val 
   private fun addOuterClassNode() {
     classNode.outerClass?.let { outerClass ->
       add(ValueNode("Outer class:", { ctx -> TypeUtils.toReadableName(outerClass, ctx.typeNameRenderMode) },
-                    searchProvider = SearchProvider.Class(outerClass)))
+                    goToProvider = GoToProvider.Class(outerClass)))
     }
   }
 
@@ -141,12 +141,12 @@ internal class ClassStructureNode(private val classNode: ClassNode, private val 
       classNode.nestHostClass?.let { nestHostClass ->
         add(ValueNode("Nest host:",
                       { ctx -> TypeUtils.toReadableName(nestHostClass, ctx.typeNameRenderMode) },
-                      searchProvider = SearchProvider.Class(nestHostClass)))
+                      goToProvider = GoToProvider.Class(nestHostClass)))
       }
 
       addTitleNodeWithElements(classNode.nestMembers, { TextNode("Nest Members") }) { _, nestMember ->
         ValueNode(displayValue = { ctx -> TypeUtils.toReadableName(nestMember, ctx.typeNameRenderMode) },
-                  searchProvider = SearchProvider.Class(nestMember))
+                  goToProvider = GoToProvider.Class(nestMember))
       }
     })
   }
@@ -155,7 +155,7 @@ internal class ClassStructureNode(private val classNode: ClassNode, private val 
     addTitleNodeWithElements(classNode.innerClasses, { TextNode("Inner classes", AllIcons.Nodes.Class) }) { _, innerClass ->
       ValueNode(displayValue = { ctx -> TypeUtils.toReadableName(innerClass.name, ctx.typeNameRenderMode) },
                 icon = AllIcons.Nodes.Class,
-                searchProvider = SearchProvider.Class(innerClass.name))
+                goToProvider = GoToProvider.Class(innerClass.name))
     }
   }
 
@@ -169,7 +169,7 @@ internal class ClassStructureNode(private val classNode: ClassNode, private val 
     addTitleNodeWithElements(classNode.interfaces, { TextNode("Interfaces", AllIcons.Nodes.Interface) }) { _, `interface` ->
       ValueNode(displayValue = { ctx -> TypeUtils.toReadableName(`interface`, ctx.typeNameRenderMode) },
                 icon = AllIcons.Nodes.Interface,
-                searchProvider = SearchProvider.Class(`interface`))
+                goToProvider = GoToProvider.Class(`interface`))
     }
   }
 
@@ -197,7 +197,7 @@ internal class ClassStructureNode(private val classNode: ClassNode, private val 
     addTitleNodeWithElements(classNode.permittedSubclasses, { TextNode("Permitted subclasses", AllIcons.General.OverridingMethod) }) { _, subclass ->
       ValueNode(displayValue = { ctx -> TypeUtils.toReadableName(subclass, ctx.typeNameRenderMode) },
                 icon = AllIcons.General.OverridingMethod,
-                searchProvider = SearchProvider.Class(subclass))
+                goToProvider = GoToProvider.Class(subclass))
     }
   }