+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/view/_internal/_structure/_common/HyperLinkNode.kt
@@ -37,6 +37,8 @@ internal open class HyperLinkNode(private val text: String, initialHyperLinkList
     return component!!
   }
 
+  override fun searchText(context: StructureTreeContext) = null
+
   fun addHyperLinkListener(hyperLinkListener: HyperLinkListener) {
     hyperLinkListeners.add(hyperLinkListener)
   }