+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Add search functionality to structure tree and constant pool (GitHub issue #3)
 
 ### Changed